// feature-1.14.0/src/types/request.d.ts
export declare namespace RequestParams {
// ...
  interface CreateAndUpdateConfigList {
    value?: string;
    key?: string;
    env: string;
    appId?: number;
    envId?: number;
    id: number | null;
    name?: string;
    remark: string;
  }
// ...
}